Remove main and fitness functions from repo
as discussed, we move these to the ludics repo
